# Conflicts:
#	yupana-api/src/main/scala/org/yupana/api/types/DataType.scala
#	yupana-core/src/main/scala/org/yupana/core/utils/CollectionUtils.scala

+++ yupana-jdbc/src/test/scala/org/yupana/jdbc/YupanaResultSetTest.scala
@@ -266,7 +266,7 @@ class YupanaResultSetTest extends FlatSpec with Matchers with MockFactory {
         DataType[Boolean],
         DataType[Int],
         DataType[String],
-        DataType[Double],
+        DataType[BigDecimal],
         DataType[Long],
         DataType[BigDecimal]
       ),
@@ -618,7 +618,7 @@ class YupanaResultSetTest extends FlatSpec with Matchers with MockFactory {
     val result = SimpleResult(
       "test",
       Seq("int", "string", "double", "time"),
-      Seq(DataType[Int], DataType[String], DataType[Double], DataType[Time]),
+      Seq(DataType[Int], DataType[String], DataType[BigDecimal], DataType[Time]),
       Iterator(
         Array[Option[Any]](Some(42), Some("foo"), None, Some(Time(1234567L)))
       )
